BUG: If the clone was canceled then a crash could occur
When the clone was canceled it would delete the repo but not set the
repo to null. This would cause the repo to be deleted for a second time when
the open occurred. This has been fixed by nulling the repo pointer after
deleting it when cloning. Also the delete has been shifted to a point
where the cloned repo has been fully clone.